18.1: Update output and restrict PyPI versions
This simplifies and cleans up the output of the command in both normal
and debug modes. It will also print out the expected and actual hash in
debug mode.

The PyPI versions have been pinned to specific versions.

The README and Snapcraft descriptions were filled out with further
detail.

This will be release 18.1 to PyPI.