iso: Add the ability to specify a mirror
This adds the ability to provide a URL to a mirror that hosts released
server and desktop ISOs. The mirror URL does not apply to any flavor or
development release ISOs. User's can find mirrors near by from the
following URL:

https://launchpad.net/ubuntu/+cdmirrors

Also cleans up the --dry-run output to actually print the link to the
ISO.