Remove embedded GPG key
The public GPG key used for Ubuntu ISO signing was embedded into the
source code to make the script transportable across installs and reduce
the dependency on the Ubuntu systems and packages.

Now the GPG is expected to be on the system or in the snap, which is the
primariy method of using this software. Therefore, it does not make
sense to have it embedded and is considered safer.