gpg: use the correct key
This uses the correct GPG key to verify ISOs.

It also fixes a known issue with the release.py file where the month was
not correctly compared when making comparisons.

It also fixes an issue when the filename was not the correct filename.
This is the case when a point release of an ISO is released.