Use the correct sha256 from SHA256SUMS file
The logic to figure out what SHA to use was wrong and matching on the
last entry with a matching flavor and arch. When it should have been
matching on the filename itself. I was already using this logic with
the mini ISO, but I cannot recall why I was doing it differently
with the other ISOs.

Fixes #6